// src/app/core/helios-api-client/model/nullableOfStatutsMembres.ts
/**
 * Helios | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const NullableOfStatutsMembres = {
    Present: 'Present',
    Suivi: 'Suivi',
    Absent: 'Absent',
    Demissionnaire: 'Demissionnaire',
    Decede: 'Decede'
} as const;
export type NullableOfStatutsMembres = typeof NullableOfStatutsMembres[keyof typeof NullableOfStatutsMembres];